problem-generator/requirements.txt
dspy-ai>=2.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0